fix: transforma as funçoes para acessar o google meet em uma classe
